transaction.go: Moved type transaction to own file.
