+++ database.go
@@ -3,7 +3,6 @@ package main
 import (
 	"database/sql"
 	_ "github.com/jackc/pgx/v4/stdlib"
-	"time"
 )
 
 var db *sql.DB
@@ -23,16 +22,6 @@ type budget struct {
 	client_id int
 	cat_id    int
 	amount    sql.NullFloat64
-}
-
-type transaction struct {
-	ID          int
-	Client_id   int
-	Cat_id      sql.NullInt64
-	Amount      float64
-	Balance     float64
-	Description sql.NullString
-	Time        time.Time
 }
 
 type stock struct {
